backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "BlackSerge",email = "sergiosaborit99@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "anyio (==4.12.0)",
    "asgiref (==3.11.0)",
    "attrs (==25.4.0)",
    "autobahn (==25.12.1)",
    "automat (==25.4.16)",
    "build (==1.3.0)",
    "cachecontrol (==0.14.4)",
    "cbor2 (==5.7.1)",
    "certifi (==2025.11.12)",
    "cffi (==2.0.0)",
    "channels (==4.3.2)",
    "charset-normalizer (==3.4.4)",
    "cleo (==2.1.0)",
    "colorama (==0.4.6)",
    "constantly (==23.10.4)",
    "crashtest (==0.4.1)",
    "cryptography (==46.0.3)",
    "daphne (==4.2.1)",
    "distlib (==0.4.0)",
    "django (==6.0)",
    "django-environ (==0.12.0)",
    "djangorestframework (==3.16.1)",
    "djangorestframework-simplejwt (==5.5.1)",
    "dulwich (==0.24.10)",
    "fastjsonschema (==2.21.2)",
    "filelock (==3.20.1)",
    "findpython (==0.7.1)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "hyperlink (==21.0.0)",
    "idna (==3.11)",
    "incremental (==24.11.0)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.3.0)",
    "keyring (==25.7.0)",
    "more-itertools (==10.8.0)",
    "msgpack (==1.1.2)",
    "packaging (==25.0)",
    "pbs-installer (==2025.12.5)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.5.1)",
    "poetry (==2.2.1)",
    "poetry-core (==2.2.1)",
    "psycopg2-binary (==2.9.11)",
    "py-ubjson (==0.16.1)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pycparser (==2.23)",
    "pyjwt (==2.10.1)",
    "pyopenssl (==25.3.0)",
    "pyproject-hooks (==1.2.0)",
    "pywin32-ctypes (==0.2.3)",
    "rapidfuzz (==3.14.3)",
    "redis (==7.1.0)",
    "requests (==2.32.5)",
    "requests-toolbelt (==1.0.0)",
    "service-identity (==24.2.0)",
    "shellingham (==1.5.4)",
    "sqlparse (==0.5.4)",
    "tomlkit (==0.13.3)",
    "trove-classifiers (==2025.12.1.14)",
    "twisted (==25.5.0)",
    "txaio (==25.12.2)",
    "typing-extensions (==4.15.0)",
    "tzdata (==2025.3)",
    "ujson (==5.11.0)",
    "urllib3 (==2.6.2)",
    "virtualenv (==20.35.4)",
    "zope-interface (==8.1.1)",
    "zstandard (==0.25.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
